+++ libs/common/src/lib/validator-constraints/is-after-1970.ts
@@ -1,8 +1,10 @@
+import { isValidDateAfter1970 } from '@ghostfolio/common/helper';
+
 import {
   ValidatorConstraint,
   ValidatorConstraintInterface
 } from 'class-validator';
-import { format, isAfter } from 'date-fns';
+import { format } from 'date-fns';
 
 @ValidatorConstraint({ name: 'isAfter1970' })
 export class IsAfter1970Constraint implements ValidatorConstraintInterface {
@@ -10,7 +12,7 @@ export class IsAfter1970Constraint implements ValidatorConstraintInterface {
     return `date must be after ${format(new Date(0), 'yyyy')}`;
   }
 
-  public validate(aDate: Date) {
-    return isAfter(aDate, new Date(0));
+  public validate(aDate: Date | string) {
+    return isValidDateAfter1970(aDate);
   }
 }