+++ libs/common/src/lib/validators/is-currency-code.ts
@@ -1,4 +1,4 @@
-import { isDerivedCurrency } from '@ghostfolio/common/helper';
+import { isValidCurrencyCode } from '@ghostfolio/common/helper';
 
 import {
   registerDecorator,
@@ -6,7 +6,6 @@ import {
   ValidatorConstraint,
   ValidatorConstraintInterface
 } from 'class-validator';
-import { isISO4217CurrencyCode } from 'class-validator';
 
 export function IsCurrencyCode(validationOptions?: ValidationOptions) {
   return function (object: object, propertyName: string) {
@@ -27,14 +26,6 @@ export class IsExtendedCurrencyConstraint implements ValidatorConstraintInterfac
   }
 
   public validate(currency: string) {
-    // Return true if currency is a derived currency or a standard ISO 4217 code
-    return (
-      isDerivedCurrency(currency) ||
-      (this.isUpperCase(currency) && isISO4217CurrencyCode(currency))
-    );
-  }
-
-  private isUpperCase(aString: string) {
-    return aString === aString?.toUpperCase();
+    return isValidCurrencyCode(currency);
   }
 }